Revert moving silence_stdio
